Split mocking into modules and functions
